+++ tests/test-text-hover.php
@@ -190,6 +190,10 @@ class Text_Hover_Test extends WP_UnitTestCase {
 		}
 	}
 
+	public function c2c_text_hover_filter_priority( $priority, $filter = '' ) {
+		return ( 'filter_20' === $filter ) ? 20 : 11;
+	}
+
 	public function capture_filter_value( $value ) {
 		return $this->captured_filter_value[ current_filter() ] = $value;
 	}
@@ -540,10 +544,10 @@ class Text_Hover_Test extends WP_UnitTestCase {
 	/**
 	 * @dataProvider get_default_filters
 	 */
-	public function test_hover_applies_to_default_filters( $filter ) {
+	public function test_hover_applies_to_default_filters( $filter, $priority = 3 ) {
 		$expected = $this->expected_text( 'coffee2code' );
 
-		$this->assertEquals( 3, has_filter( $filter, array( c2c_TextHover::get_instance(), 'text_hover' ) ) );
+		$this->assertEquals( $priority, has_filter( $filter, array( c2c_TextHover::get_instance(), 'text_hover' ) ) );
 		$this->assertGreaterThan( 0, strpos( apply_filters( $filter, 'a coffee2code' ), $expected ) );
 	}
 
@@ -605,6 +609,26 @@ class Text_Hover_Test extends WP_UnitTestCase {
 		c2c_TextHover::get_instance()->register_filters(); // Plugins would typically register their filter before this originally fires
 
 		$this->assertEquals( $this->expected_text( 'coffee2code' ), apply_filters( 'custom_filter', 'coffee2code' ) );
+	}
+
+	/*
+	 * filter: c2c_text_hover_filter_priority
+	 */
+
+	public function test_changing_priority_via_c2c_text_hover_filter_priority() {
+		$filters = $this->get_filter_names();
+
+		$this->unhook_default_filters();
+
+		add_filter( 'c2c_text_hover_filter_priority', array( $this, 'c2c_text_hover_filter_priority' ) );
+
+		c2c_TextHover::get_instance()->register_filters(); // Plugins would typically register their filter before this originally fires
+
+		$priority = 11;
+
+		foreach ( $filters as $filter ) {
+			$this->test_hover_applies_to_default_filters( $filter, $priority );
+		}
 	}
 
 	/*